+++ sdk/src/main/java/com/wonderpush/sdk/remoteconfig/AsyncHttpClientRemoteConfigFetcher.java
@@ -39,7 +39,7 @@ public class AsyncHttpClientRemoteConfigFetcher implements RemoteConfigFetcher {
                         return;
                     }
 
-                    String version = response.optString("version", Integer.toString(response.optInt("version", 0)));
+                    String version = response.optString("version", Long.toString(response.optLong("version", 0)));
 
                     long maxAge = response.optLong("maxAge", Constants.REMOTE_CONFIG_DEFAULT_MAXIMUM_CONFIG_AGE);
                     if (version != null) {
